refactor: rename `CreateWallet` steps by context and split tests per step
