+++ src/domains/wallet/pages/CreateWallet/ConfirmPassphrase.test.tsx
@@ -1,0 +1,40 @@
+/* eslint-disable @typescript-eslint/require-await */
+import { Profile } from "@arkecosystem/platform-sdk-profiles";
+import { renderHook } from "@testing-library/react-hooks";
+import React from "react";
+import { FormProvider, useForm } from "react-hook-form";
+import { env, getDefaultProfileId, render } from "utils/testing-library";
+
+import { ConfirmPassphraseStep } from "./ConfirmPassphraseStep";
+
+let profile: Profile;
+
+describe("CreateWallet - Step 3", () => {
+	beforeEach(() => {
+		profile = env.profiles().findById(getDefaultProfileId());
+
+		for (const wallet of profile.wallets().values()) {
+			profile.wallets().forget(wallet.id());
+		}
+	});
+
+	it("should render 3rd step", () => {
+		const { result: form } = renderHook(() =>
+			useForm({
+				defaultValues: {
+					mnemonic: "hamster giggle left flush sock appear mule either order solve spirit neutral",
+				},
+			}),
+		);
+		const { getByTestId, getAllByTestId } = render(
+			<FormProvider {...form.current}>
+				<ConfirmPassphraseStep />
+			</FormProvider>,
+		);
+
+		expect(getByTestId("CreateWallet__third-step")).toBeTruthy();
+		expect(getAllByTestId("MnemonicVerificationOptions__button").length).toBeGreaterThan(1);
+
+		expect(form.current.getValues()).toEqual({ verification: undefined });
+	});
+});
