+++ src/domains/wallet/pages/CreateWallet/CreateWallet.test.tsx
@@ -1,50 +1,40 @@
 /* eslint-disable @typescript-eslint/require-await */
 import { BIP39 } from "@arkecosystem/platform-sdk-crypto";
-import { Profile, ProfileSetting } from "@arkecosystem/platform-sdk-profiles";
-import { act, renderHook } from "@testing-library/react-hooks";
-import { toasts } from "app/services";
-import { availableNetworksMock } from "domains/network/data";
+import { Profile } from "@arkecosystem/platform-sdk-profiles";
+import { act } from "@testing-library/react-hooks";
 import { translations as walletTranslations } from "domains/wallet/i18n";
-import electron from "electron";
 import { createMemoryHistory } from "history";
 import React from "react";
-import { FormProvider, useForm } from "react-hook-form";
 import { Route } from "react-router-dom";
 import {
 	act as actAsync,
 	env,
 	fireEvent,
 	getDefaultProfileId,
-	render,
 	renderWithRouter,
 	screen,
 	waitFor,
 } from "utils/testing-library";
 
 import { CreateWallet } from "./CreateWallet";
-import { FirstStep } from "./Step1";
-import { SecondStep } from "./Step2";
-import { ThirdStep } from "./Step3";
-import { FourthStep } from "./Step4";
 
 jest.setTimeout(8000);
 
 let profile: Profile;
 let bip39GenerateMock: any;
 
+const fixtureProfileId = getDefaultProfileId();
 const passphrase = "power return attend drink piece found tragic fire liar page disease combine";
 
-beforeAll(() => {
-	bip39GenerateMock = jest.spyOn(BIP39, "generate").mockReturnValue(passphrase);
-});
-
-const fixtureProfileId = getDefaultProfileId();
-
-afterAll(() => {
-	bip39GenerateMock.mockRestore();
-});
-
 describe("CreateWallet", () => {
+	beforeAll(() => {
+		bip39GenerateMock = jest.spyOn(BIP39, "generate").mockReturnValue(passphrase);
+	});
+
+	afterAll(() => {
+		bip39GenerateMock.mockRestore();
+	});
+
 	beforeEach(() => {
 		profile = env.profiles().findById(fixtureProfileId);
 
@@ -57,240 +47,6 @@ describe("CreateWallet", () => {
 
 	afterEach(() => {
 		bip39GenerateMock.mockRestore();
-	});
-
-	describe("1st step", () => {
-		it("should render", () => {
-			const { result: form } = renderHook(() => useForm());
-			const { getByTestId, asFragment } = render(
-				<FormProvider {...form.current}>
-					<FirstStep env={env} profile={profile} />
-				</FormProvider>,
-			);
-
-			expect(getByTestId("CreateWallet__first-step")).toBeTruthy();
-			expect(asFragment()).toMatchSnapshot();
-
-			const selectNetworkInput = getByTestId("SelectNetworkInput__input");
-			expect(selectNetworkInput).toBeTruthy();
-		});
-
-		it("should render without test networks", async () => {
-			profile.settings().set(ProfileSetting.UseTestNetworks, false);
-
-			const { result: form } = renderHook(() => useForm());
-			const { getByTestId, asFragment, queryByTestId } = render(
-				<FormProvider {...form.current}>
-					<FirstStep env={env} profile={profile} />
-				</FormProvider>,
-			);
-
-			expect(getByTestId("CreateWallet__first-step")).toBeTruthy();
-
-			const selectNetworkInput = getByTestId("SelectNetworkInput__input");
-			expect(selectNetworkInput).toBeTruthy();
-
-			act(() => {
-				fireEvent.focus(selectNetworkInput);
-			});
-
-			expect(queryByTestId("NetworkIcon-ARK-ark.mainnet")).toBeInTheDocument();
-			expect(queryByTestId("NetworkIcon-ARK-ark.devnet")).toBeNull();
-
-			expect(asFragment()).toMatchSnapshot();
-
-			profile.settings().set(ProfileSetting.UseTestNetworks, true);
-		});
-	});
-
-	describe("2nd step", () => {
-		it("should render", async () => {
-			const { result: form } = renderHook(() =>
-				useForm({
-					defaultValues: {
-						mnemonic: "test mnemonic",
-						wallet: {
-							address: () => "address",
-						},
-					},
-				}),
-			);
-
-			jest.spyOn(electron.remote.dialog, "showSaveDialog").mockImplementation(() => ({
-				filePath: "filePath",
-			}));
-
-			const { getByTestId, asFragment } = render(
-				<FormProvider {...form.current}>
-					<SecondStep />
-				</FormProvider>,
-			);
-
-			expect(asFragment()).toMatchSnapshot();
-
-			const writeTextMock = jest.fn();
-			const clipboardOriginal = navigator.clipboard;
-			// @ts-ignore
-			navigator.clipboard = { writeText: writeTextMock };
-
-			act(() => {
-				fireEvent.click(getByTestId("CreateWallet__copy"));
-			});
-
-			await waitFor(() => expect(writeTextMock).toHaveBeenCalledWith("test mnemonic"));
-
-			// @ts-ignore
-			navigator.clipboard = clipboardOriginal;
-		});
-
-		it("should show success toast on succesfull download", async () => {
-			const { result: form } = renderHook(() =>
-				useForm({
-					defaultValues: {
-						mnemonic: "test mnemonic",
-						wallet: {
-							address: () => "address",
-						},
-					},
-				}),
-			);
-
-			jest.spyOn(electron.remote.dialog, "showSaveDialog").mockImplementation(() => ({
-				filePath: "filePath",
-			}));
-
-			const toastSpy = jest.spyOn(toasts, "success");
-
-			const { getByTestId, asFragment } = render(
-				<FormProvider {...form.current}>
-					<SecondStep />
-				</FormProvider>,
-			);
-
-			await act(async () => {
-				fireEvent.click(getByTestId("CreateWallet__download"));
-			});
-
-			expect(toastSpy).toHaveBeenCalled();
-			toastSpy.mockRestore();
-		});
-
-		it("should not show success toast on cancelled download", async () => {
-			const { result: form } = renderHook(() =>
-				useForm({
-					defaultValues: {
-						mnemonic: "test mnemonic",
-						wallet: {
-							address: () => "address",
-						},
-					},
-				}),
-			);
-
-			jest.spyOn(electron.remote.dialog, "showSaveDialog").mockImplementation(() => ({
-				filePath: undefined,
-			}));
-
-			const toastSpy = jest.spyOn(toasts, "success");
-
-			const { getByTestId, asFragment } = render(
-				<FormProvider {...form.current}>
-					<SecondStep />
-				</FormProvider>,
-			);
-
-			await act(async () => {
-				fireEvent.click(getByTestId("CreateWallet__download"));
-			});
-
-			expect(toastSpy).not.toHaveBeenCalled();
-			toastSpy.mockRestore();
-		});
-
-		it("should show error toast on error", async () => {
-			const { result: form } = renderHook(() =>
-				useForm({
-					defaultValues: {
-						mnemonic: "test mnemonic",
-						wallet: {
-							address: () => "address",
-						},
-					},
-				}),
-			);
-
-			jest.spyOn(electron.remote.dialog, "showSaveDialog").mockImplementation(() => {
-				throw new Error("Error");
-			});
-
-			const toastSpy = jest.spyOn(toasts, "error");
-
-			const { getByTestId, asFragment } = render(
-				<FormProvider {...form.current}>
-					<SecondStep />
-				</FormProvider>,
-			);
-
-			await act(async () => {
-				fireEvent.click(getByTestId("CreateWallet__download"));
-			});
-
-			expect(toastSpy).toHaveBeenCalled();
-			toastSpy.mockRestore();
-		});
-	});
-
-	it("should render 3rd step", () => {
-		const { result: form } = renderHook(() =>
-			useForm({
-				defaultValues: {
-					mnemonic: "hamster giggle left flush sock appear mule either order solve spirit neutral",
-				},
-			}),
-		);
-		const { getByTestId, getAllByTestId } = render(
-			<FormProvider {...form.current}>
-				<ThirdStep />
-			</FormProvider>,
-		);
-
-		expect(getByTestId("CreateWallet__third-step")).toBeTruthy();
-		expect(getAllByTestId("MnemonicVerificationOptions__button").length).toBeGreaterThan(1);
-
-		expect(form.current.getValues()).toEqual({ verification: undefined });
-	});
-
-	it("should render 4th step", async () => {
-		const { result: form } = renderHook(() =>
-			useForm({
-				defaultValues: {
-					network: availableNetworksMock[1],
-					wallet: {
-						address: () => "TEST-WALLET-ADDRESS",
-					},
-				},
-			}),
-		);
-
-		const { asFragment, getByTestId, getByText } = render(
-			<FormProvider {...form.current}>
-				<FourthStep nameMaxLength={42} />
-			</FormProvider>,
-		);
-
-		expect(getByTestId("CreateWallet__fourth-step")).toBeTruthy();
-		expect(asFragment()).toMatchSnapshot();
-
-		expect(getByText("ARK Devnet")).toBeTruthy();
-		expect(getByText("TEST-WALLET-ADDRESS")).toBeTruthy();
-
-		const walletNameInput = getByTestId("CreateWallet__wallet-name");
-
-		await act(async () => {
-			fireEvent.change(walletNameInput, { target: { value: "Test" } });
-		});
-
-		expect(form.current.getValues()).toEqual({ name: "Test" });
 	});
 
 	it.each([
@@ -439,262 +195,6 @@ describe("CreateWallet", () => {
 		expect(asFragment()).toMatchSnapshot();
 
 		historySpy.mockRestore();
-	});
-
-	it("should create a wallet and use encryption password", async () => {
-		const history = createMemoryHistory();
-		const createURL = `/profiles/${fixtureProfileId}/wallets/create`;
-		history.push(createURL);
-
-		const historySpy = jest.spyOn(history, "push").mockImplementation();
-
-		const { queryAllByText, getByTestId, getByText, asFragment, getAllByTestId } = renderWithRouter(
-			<Route path="/profiles/:profileId/wallets/create">
-				<CreateWallet />
-			</Route>,
-			{
-				routes: [createURL],
-				history,
-			},
-		);
-
-		await waitFor(() => expect(getByTestId("CreateWallet__first-step")).toBeTruthy());
-		expect(asFragment()).toMatchSnapshot();
-
-		const selectNetworkInput = getByTestId("SelectNetworkInput__input");
-		const continueButton = getByTestId("CreateWallet__continue-button");
-		const backButton = getByTestId("CreateWallet__back-button");
-
-		await actAsync(async () => {
-			fireEvent.change(selectNetworkInput, { target: { value: "Ark Dev" } });
-		});
-
-		await actAsync(async () => {
-			fireEvent.keyDown(selectNetworkInput, { key: "Enter", code: 13 });
-		});
-
-		await waitFor(() => expect(selectNetworkInput).toHaveValue("ARK Devnet"));
-		await waitFor(() => expect(continueButton).not.toHaveAttribute("disabled"));
-
-		await actAsync(async () => {
-			fireEvent.click(getByTestId("CreateWallet__continue-button"));
-		});
-
-		await waitFor(() => expect(profile.wallets().values().length).toBe(1));
-
-		await waitFor(() => expect(getByTestId("CreateWallet__second-step")).toBeTruthy());
-
-		await actAsync(async () => {
-			fireEvent.click(getByTestId("CreateWallet__continue-button"));
-		});
-
-		await waitFor(() => expect(getByTestId("CreateWallet__third-step")).toBeTruthy());
-
-		await actAsync(async () => {
-			fireEvent.click(backButton);
-		});
-
-		await waitFor(() => expect(getByTestId("CreateWallet__second-step")).toBeTruthy());
-
-		await actAsync(async () => {
-			fireEvent.click(continueButton);
-		});
-
-		await waitFor(() => expect(getByTestId("CreateWallet__third-step")).toBeTruthy());
-
-		const walletMnemonic = passphrase.split(" ");
-		for (let i = 0; i < 3; i++) {
-			const wordNumber = parseInt(getByText(/Select the/).innerHTML.replace(/Select the/, ""));
-
-			await actAsync(async () => {
-				fireEvent.click(getByText(walletMnemonic[wordNumber - 1]));
-				if (i < 2) {
-					await waitFor(() => expect(queryAllByText(/The #([0-9]+) word/).length === 2 - i));
-				}
-			});
-		}
-		await waitFor(() => expect(continueButton).not.toHaveAttribute("disabled"));
-
-		await actAsync(async () => {
-			fireEvent.click(continueButton);
-		});
-
-		await waitFor(() => expect(getByTestId("EncryptPassword")).toBeTruthy());
-
-		await actAsync(async () => {
-			fireEvent.click(getByTestId("CreateWallet__skip-button"));
-		});
-
-		await waitFor(() => expect(getByTestId("CreateWallet__fourth-step")).toBeTruthy());
-
-		await actAsync(async () => {
-			fireEvent.click(backButton);
-		});
-
-		await waitFor(() => expect(getByTestId("EncryptPassword")).toBeTruthy());
-
-		await actAsync(async () => {
-			fireEvent.click(getByTestId("CreateWallet__skip-button"));
-		});
-
-		await waitFor(() => expect(getByTestId("CreateWallet__fourth-step")).toBeTruthy());
-
-		await actAsync(async () => {
-			fireEvent.click(backButton);
-		});
-
-		await waitFor(() => expect(getByTestId("EncryptPassword")).toBeTruthy());
-
-		const passwordInput = getAllByTestId("Input")[0];
-		const confirmPassword = getAllByTestId("Input")[1];
-
-		await actAsync(async () => {
-			fireEvent.input(passwordInput, { target: { value: "password" } });
-		});
-
-		await actAsync(async () => {
-			fireEvent.input(confirmPassword, { target: { value: "password" } });
-		});
-
-		await waitFor(() => expect(confirmPassword).toHaveValue("password"));
-
-		act(() => {
-			fireEvent.click(getByTestId("CreateWallet__continue-button"));
-		});
-
-		await waitFor(() => expect(getByTestId("CreateWallet__fourth-step")).toBeTruthy());
-
-		act(() => {
-			fireEvent.click(getByTestId("CreateWallet__save-button"));
-		});
-
-		await waitFor(() => expect(historySpy).toHaveBeenCalled());
-
-		expect(asFragment()).toMatchSnapshot();
-
-		historySpy.mockRestore();
-	});
-
-	it("should fail creating a wallet with encryption password", async () => {
-		const history = createMemoryHistory();
-		const createURL = `/profiles/${fixtureProfileId}/wallets/create`;
-
-		const walletSpy = jest
-			.spyOn(profile.wallets(), "importByMnemonicWithEncryption")
-			.mockRejectedValue(new Error("failed"));
-
-		history.push(createURL);
-
-		const { queryAllByText, getByTestId, getByText, asFragment, getAllByTestId } = renderWithRouter(
-			<Route path="/profiles/:profileId/wallets/create">
-				<CreateWallet />
-			</Route>,
-			{
-				routes: [createURL],
-				history,
-			},
-		);
-
-		await waitFor(() => expect(getByTestId("CreateWallet__first-step")).toBeTruthy());
-		expect(asFragment()).toMatchSnapshot();
-
-		const selectNetworkInput = getByTestId("SelectNetworkInput__input");
-		const continueButton = getByTestId("CreateWallet__continue-button");
-		const backButton = getByTestId("CreateWallet__back-button");
-
-		await actAsync(async () => {
-			fireEvent.change(selectNetworkInput, { target: { value: "Ark Dev" } });
-		});
-
-		await actAsync(async () => {
-			fireEvent.keyDown(selectNetworkInput, { key: "Enter", code: 13 });
-		});
-
-		await waitFor(() => expect(selectNetworkInput).toHaveValue("ARK Devnet"));
-		await waitFor(() => expect(continueButton).not.toHaveAttribute("disabled"));
-
-		await actAsync(async () => {
-			fireEvent.click(getByTestId("CreateWallet__continue-button"));
-		});
-
-		await waitFor(() => expect(profile.wallets().values().length).toBe(1));
-
-		await waitFor(() => expect(getByTestId("CreateWallet__second-step")).toBeTruthy());
-
-		await actAsync(async () => {
-			fireEvent.click(getByTestId("CreateWallet__continue-button"));
-		});
-
-		await waitFor(() => expect(getByTestId("CreateWallet__third-step")).toBeTruthy());
-
-		await actAsync(async () => {
-			fireEvent.click(backButton);
-		});
-
-		await waitFor(() => expect(getByTestId("CreateWallet__second-step")).toBeTruthy());
-
-		await actAsync(async () => {
-			fireEvent.click(continueButton);
-		});
-
-		await waitFor(() => expect(getByTestId("CreateWallet__third-step")).toBeTruthy());
-
-		const walletMnemonic = passphrase.split(" ");
-		for (let i = 0; i < 3; i++) {
-			const wordNumber = parseInt(getByText(/Select the/).innerHTML.replace(/Select the/, ""));
-
-			await actAsync(async () => {
-				fireEvent.click(getByText(walletMnemonic[wordNumber - 1]));
-				if (i < 2) {
-					await waitFor(() => expect(queryAllByText(/The #([0-9]+) word/).length === 2 - i));
-				}
-			});
-		}
-		await waitFor(() => expect(continueButton).not.toHaveAttribute("disabled"));
-
-		await actAsync(async () => {
-			fireEvent.click(continueButton);
-		});
-
-		await waitFor(() => expect(getByTestId("EncryptPassword")).toBeTruthy());
-
-		await actAsync(async () => {
-			fireEvent.click(getByTestId("CreateWallet__skip-button"));
-		});
-
-		await waitFor(() => expect(getByTestId("CreateWallet__fourth-step")).toBeTruthy());
-
-		await actAsync(async () => {
-			fireEvent.click(backButton);
-		});
-
-		await waitFor(() => expect(getByTestId("EncryptPassword")).toBeTruthy());
-
-		const passwordInput = getAllByTestId("Input")[0];
-		const confirmPassword = getAllByTestId("Input")[1];
-
-		await actAsync(async () => {
-			fireEvent.input(passwordInput, { target: { value: "password" } });
-		});
-
-		await actAsync(async () => {
-			fireEvent.input(confirmPassword, { target: { value: "password" } });
-		});
-
-		await waitFor(() => expect(confirmPassword).toHaveValue("password"));
-
-		await actAsync(async () => {
-			fireEvent.click(getByTestId("CreateWallet__continue-button"));
-		});
-
-		await waitFor(() => expect(getByTestId("CreateWallet__fourth-step")).toBeTruthy());
-
-		await actAsync(async () => {
-			fireEvent.click(getByTestId("CreateWallet__save-button"));
-		});
-
-		await waitFor(() => expect(walletSpy).toHaveBeenCalled());
-		walletSpy.mockRestore();
 	});
 
 	it("should not have a pending wallet if leaving on step 1", async () => {
