+++ src/domains/wallet/pages/CreateWallet/SuccessStep.tsx
@@ -10,7 +10,7 @@ import React, { useState } from "react";
 import { useFormContext } from "react-hook-form";
 import { useTranslation } from "react-i18next";
 
-export const FourthStep = ({ nameMaxLength, profile }: { nameMaxLength: number; profile: Profile }) => {
+export const SuccessStep = ({ nameMaxLength, profile }: { nameMaxLength: number; profile: Profile }) => {
 	const { getValues, register, watch } = useFormContext();
 
 	// getValues does not get the value of `defaultValues` on first render
